Added helper function to avoid code duplication when reducing size of services. Also fixed all bugs with base64 conversion and downloading on frontend and after first deploying Xnode.
